+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.1.0 - 2019-02-27
+
+- dump the version of the compiler/interpreter in verbose mode
+- fix gcc 4 (hello.c): add missing return value
+- update obsolete dmd options
+
 ## 1.0.0 - 2019-02-21
 
-- gemify
+- convert the script to a gem and release it
 - add tests for Deno, GraalVM JavaScript, Nim, and TruffleRuby
 - add --json option to render the results as a JSON array
 - add changelog